изменение типа

+++ src/main/java/IncomeService.java
@@ -1,6 +1,6 @@
 public class IncomeService {
-        double calculate(double sum){
-            int percent = 3;
+        double calculate(int sum){
+            double percent = 3;
             double income = sum * percent / 100;
 
             int limit = 3_000;
